+++ components/goal-hierarchy/index.ts
@@ -1,0 +1,3 @@
+export { GoalHierarchyView } from './GoalHierarchyView';
+export { GoalLevelBlock } from './GoalLevelBlock';
+export { DayColumn } from './DayColumn';